LayoutTests/editing/selection/selection-change-in-mutation-event-by-remove-children-expected.txt

PASS selection.anchorNode is sample
PASS selection.anchorOffset is 0
PASS selection.focusNode is sample
PASS selection.focusOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
